Split off AssetTab components
